Changed folder name to Manager to test and got all time pi data working!
